% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name = "Nick", punctuation = "!")
}
\arguments{
\item{name}{The name of the person to praise}

\item{punctuation}{The tone of the praise}
}
\description{
Let's deliver some praise!
}
\examples{

}
