% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Deliver Praise}
\usage{
praise(name = "Nick", adjective = "the best", punctuation = "!")
}
\arguments{
\item{name}{The name of the person to praise}

\item{adjective}{describe the tone of the praise for this specific sentence structure.}

\item{punctuation}{The tone of the praise}
}
\description{
Let's deliver some praise!
}
\examples{
praise(name = "Miles",
       adjective = "amazing",
       punctuation = "!")
}
